Extract predicted components from a Random Planted Forest

Prediction components are a functional decomposition of the model prediction. The sum of all components equals the overall predicted value for an observation.

Usage

extract_components(object, new_data, max_interaction = NULL, predictors = NULL)

Arguments

object

A fit object of class rpf.

new_data

Data for new observations to predict.

max_interaction

integer or NULL: Maximum degree of interactions to consider. Default will use the max_interaction parameter from the rpf object. Must be between 1 (main effects only) and the max_interaction of the rpf object.

predictors

character or NULL: Vector of one or more column names of predictor variables in new_data to extract components for. If NULL, all variables and their interactions are returned.

Value

A data.table with the same number of rows as new_data and one column for each main and interaction term requested, including the intercept. For multiclass classification, the number of output columns is multiplied by the number of levels in the outcome.

Details

Extracts all possible components up to max_interaction degrees, up to the value set when calling rpf(). The intercept is always included. Optionally predictors can be specified to only include components including the given variables. If max_interaction is greater than length(predictors), the max_interaction will be lowered accordingly.

Note

Depending on the number of predictors and max_interaction, the number of components will increase drastically to sum(choose(ncol(new_data), seq_len(max_interaction))).

Examples

# Regression task, only some predictors
train <-  mtcars[1:20, 1:4]
test <-  mtcars[21:32, 1:4]

set.seed(23)
rpfit <- rpf(mpg ~ ., data = train, max_interaction = 3, ntrees = 30)

# Extract all components, including main effects and interaction terms up to `max_interaction`
(components <- extract_components(rpfit, test))
#> $m
#>     intercept        cyl       disp         hp  cyl:disp      cyl:hp
#>  1:  20.00087  1.3768855  2.6463328  1.7329858 0.1226847  0.10609368
#>  2:  20.00087 -0.7992901 -0.8268513 -0.0064323 0.1684562 -0.09246586
#>  3:  20.00087 -0.7992901 -0.8268513 -0.0064323 0.1684562 -0.09246586
#>  4:  20.00087 -0.7992901 -0.8268513 -2.9467399 0.1684562 -0.07370080
#>  5:  20.00087 -0.7992901 -1.3844161  0.1184759 0.1753777 -0.09246575
#>  6:  20.00087  1.3768855  4.2550639  4.5810324 0.4621982 -0.27912407
#>  7:  20.00087  1.3768855  2.6463328  4.5810324 0.1226847 -0.27912407
#>  8:  20.00087  1.3768855  4.2550639  0.5340290 0.4621982  0.05269070
#>  9:  20.00087 -0.7992901 -0.8268513 -2.9467399 0.1684562 -0.07370080
#> 10:  20.00087 -0.4055283  1.0484424  0.1184759 0.0909271 -0.03448583
#> 11:  20.00087 -0.7992901 -0.8268513 -2.9467399 0.1684562 -0.07370080
#> 12:  20.00087  1.3768855  2.6463328  0.1584594 0.1226847  0.12780335
#>         disp:hp  cyl:disp:hp
#>  1: -0.31696214 -0.068887656
#>  2:  0.08826036 -0.033709443
#>  3:  0.08826036 -0.033709443
#>  4: -0.27259694  0.004351475
#>  5:  0.12779098 -0.031006151
#>  6: -0.17193448  0.016794219
#>  7: -0.46815054  0.097934960
#>  8: -0.22361410 -0.032020510
#>  9: -0.27259694  0.004351475
#> 10: -0.28526023  0.024649550
#> 11: -0.27259694  0.004351475
#> 12: -0.08930346 -0.079140085
#> 
#> $x
#>     cyl  disp  hp
#>  1:   4 120.1  97
#>  2:   8 318.0 150
#>  3:   8 304.0 150
#>  4:   8 350.0 245
#>  5:   8 400.0 175
#>  6:   4  79.0  66
#>  7:   4 120.3  91
#>  8:   4  95.1 113
#>  9:   8 351.0 264
#> 10:   6 145.0 175
#> 11:   8 301.0 335
#> 12:   4 121.0 109
#> 
#> attr(,"class")
#> [1] "rpf_components" "list"          

# sums to prediction
cbind(rowSums(components$m), predict(rpfit, test))
#>    rowSums(components$m)    .pred
#> 1               25.60000 25.60000
#> 2               18.49883 18.49883
#> 3               18.49883 18.49883
#> 4               15.25450 15.25450
#> 5               18.11533 18.11533
#> 6               30.24178 30.24178
#> 7               28.07846 28.07846
#> 8               26.42610 26.42610
#> 9               15.25450 15.25450
#> 10              20.55809 20.55809
#> 11              15.25450 15.25450
#> 12              24.26459 24.26459

# Only get components with interactions of a lower degree, ignoring 3-way interactions
extract_components(rpfit, test, max_interaction = 2)
#> $m
#>     intercept        cyl       disp         hp  cyl:disp      cyl:hp
#>  1:  20.00087  1.3768855  2.6463328  1.7329858 0.1226847  0.10609368
#>  2:  20.00087 -0.7992901 -0.8268513 -0.0064323 0.1684562 -0.09246586
#>  3:  20.00087 -0.7992901 -0.8268513 -0.0064323 0.1684562 -0.09246586
#>  4:  20.00087 -0.7992901 -0.8268513 -2.9467399 0.1684562 -0.07370080
#>  5:  20.00087 -0.7992901 -1.3844161  0.1184759 0.1753777 -0.09246575
#>  6:  20.00087  1.3768855  4.2550639  4.5810324 0.4621982 -0.27912407
#>  7:  20.00087  1.3768855  2.6463328  4.5810324 0.1226847 -0.27912407
#>  8:  20.00087  1.3768855  4.2550639  0.5340290 0.4621982  0.05269070
#>  9:  20.00087 -0.7992901 -0.8268513 -2.9467399 0.1684562 -0.07370080
#> 10:  20.00087 -0.4055283  1.0484424  0.1184759 0.0909271 -0.03448583
#> 11:  20.00087 -0.7992901 -0.8268513 -2.9467399 0.1684562 -0.07370080
#> 12:  20.00087  1.3768855  2.6463328  0.1584594 0.1226847  0.12780335
#>         disp:hp
#>  1: -0.31696214
#>  2:  0.08826036
#>  3:  0.08826036
#>  4: -0.27259694
#>  5:  0.12779098
#>  6: -0.17193448
#>  7: -0.46815054
#>  8: -0.22361410
#>  9: -0.27259694
#> 10: -0.28526023
#> 11: -0.27259694
#> 12: -0.08930346
#> 
#> $x
#>     cyl  disp  hp
#>  1:   4 120.1  97
#>  2:   8 318.0 150
#>  3:   8 304.0 150
#>  4:   8 350.0 245
#>  5:   8 400.0 175
#>  6:   4  79.0  66
#>  7:   4 120.3  91
#>  8:   4  95.1 113
#>  9:   8 351.0 264
#> 10:   6 145.0 175
#> 11:   8 301.0 335
#> 12:   4 121.0 109
#> 
#> attr(,"class")
#> [1] "rpf_components" "list"          

# Only retrieve main effects
(main_effects <- extract_components(rpfit, test, max_interaction = 1))
#> $m
#>     intercept        cyl       disp         hp
#>  1:  20.00087  1.3768855  2.6463328  1.7329858
#>  2:  20.00087 -0.7992901 -0.8268513 -0.0064323
#>  3:  20.00087 -0.7992901 -0.8268513 -0.0064323
#>  4:  20.00087 -0.7992901 -0.8268513 -2.9467399
#>  5:  20.00087 -0.7992901 -1.3844161  0.1184759
#>  6:  20.00087  1.3768855  4.2550639  4.5810324
#>  7:  20.00087  1.3768855  2.6463328  4.5810324
#>  8:  20.00087  1.3768855  4.2550639  0.5340290
#>  9:  20.00087 -0.7992901 -0.8268513 -2.9467399
#> 10:  20.00087 -0.4055283  1.0484424  0.1184759
#> 11:  20.00087 -0.7992901 -0.8268513 -2.9467399
#> 12:  20.00087  1.3768855  2.6463328  0.1584594
#> 
#> $x
#>     cyl  disp  hp
#>  1:   4 120.1  97
#>  2:   8 318.0 150
#>  3:   8 304.0 150
#>  4:   8 350.0 245
#>  5:   8 400.0 175
#>  6:   4  79.0  66
#>  7:   4 120.3  91
#>  8:   4  95.1 113
#>  9:   8 351.0 264
#> 10:   6 145.0 175
#> 11:   8 301.0 335
#> 12:   4 121.0 109
#> 
#> attr(,"class")
#> [1] "rpf_components" "list"          

# The difference is the combined contribution of interaction effects
cbind(rowSums(main_effects$m), predict(rpfit, test))
#>    rowSums(main_effects$m)    .pred
#> 1                 25.75707 25.60000
#> 2                 18.36829 18.49883
#> 3                 18.36829 18.49883
#> 4                 15.42799 15.25450
#> 5                 17.93564 18.11533
#> 6                 30.21385 30.24178
#> 7                 28.60512 28.07846
#> 8                 26.16685 26.42610
#> 9                 15.42799 15.25450
#> 10                20.76226 20.55809
#> 11                15.42799 15.25450
#> 12                24.18255 24.26459