randomPlantedForest: Random Planted Forest: A Directly Interpretable Tree Ensemble

An implementation of the Random Planted Forest algorithm for directly interpretable tree ensembles based on a functional ANOVA decomposition.

See also

Author

Maintainer: Lukas Burk burk@leibniz-bips.de (ORCID)

Authors:

Joseph Theo Meyer

Munir Hiabu

Maike Spankus

Marvin N. Wright cran@wrig.de