Extract predicted components from a Random Planted Forest

Prediction components are a functional decomposition of the model prediction. The sum of all components equals the overall predicted value for an observation.

Usage

predict_components(object, new_data, max_interaction = NULL, predictors = NULL)

Arguments

object

A fit object of class rpf.

new_data

Data for new observations to predict.

max_interaction

integer or NULL: Maximum degree of interactions to consider. Default will use the max_interaction parameter from the rpf object. Must be between 1 (main effects only) and the max_interaction of the rpf object.

predictors

character or NULL: Vector of one or more column names of predictor variables in new_data to extract components for. If NULL, all variables and their interactions are returned.

Value

A list with elements:

m (data.table): Components for each main effect and interaction term, representing the functional decomposition of the prediction. All components together with the intercept sum up to the prediction. For multiclass classification, the number of output columns is multiplied by the number of levels in the outcome.

intercept (numeric(1)): Expected value of the prediction.

x (data.table): Copy of new_data containing predictors selected by predictors.

target_levels (character): For multiclass classification only: Vector of target levels which can be used to disassemble m, as names include both term and target level.

Details

Extracts all possible components up to max_interaction degrees, up to the value set when calling rpf(). The intercept is always included. Optionally predictors can be specified to only include components including the given variables. If max_interaction is greater than length(predictors), the max_interaction will be lowered accordingly.

Note

Depending on the number of predictors and max_interaction, the number of components will increase drastically to sum(choose(ncol(new_data), seq_len(max_interaction))).

Examples

# Regression task, only some predictors
train <-  mtcars[1:20, 1:4]
test <-  mtcars[21:32, 1:4]

set.seed(23)
rpfit <- rpf(mpg ~ ., data = train, max_interaction = 3, ntrees = 30)

# Extract all components, including main effects and interaction terms up to `max_interaction`
(components <- predict_components(rpfit, test))
#> $m
#>            cyl      disp         hp    cyl:disp       cyl:hp     disp:hp
#>  1:  1.3770452  2.603519  0.8918936  0.23153106 -0.267467990 -0.16388358
#>  2: -0.7695683 -1.465990 -0.1106307  0.01146022  0.006589163  0.23806033
#>  3: -0.7695683 -1.465990 -0.1106307  0.01146022  0.006589163  0.23806033
#>  4: -0.7695683 -1.465990 -2.8044495  0.01146022 -0.105503856 -0.26797812
#>  5: -0.7695683 -1.334206 -0.3514101  0.01096113 -0.010321204  0.42309626
#>  6:  1.3770452  4.125237  3.7886809  0.46853941 -0.002318937  0.94645152
#>  7:  1.3770452  2.603519  3.7886809  0.23153106 -0.002318937  0.31230241
#>  8:  1.3770452  4.125237  0.4823373  0.46853941 -0.205562593 -0.19865944
#>  9: -0.7695683 -1.465990 -2.8044495  0.01146022 -0.105503856 -0.26797812
#> 10: -0.4832547  1.694664 -0.3514101 -0.16986248  0.127061572 -0.11840583
#> 11: -0.7695683 -1.465990 -2.8044495  0.01146022 -0.105503856 -0.26797812
#> 12:  1.3770452  2.603519  0.4516721  0.23153106 -0.200179679 -0.05174859
#>     cyl:disp:hp
#>  1:  0.07242737
#>  2: -0.04685440
#>  3: -0.04685440
#>  4:  0.13144921
#>  5: -0.01982741
#>  6: -0.10343096
#>  7: -0.03536481
#>  8: -0.05597535
#>  9:  0.13144921
#> 10: -0.09596553
#> 11:  0.13144921
#> 12:  0.05100935
#> 
#> $intercept
#> [1] 20.17863
#> 
#> $x
#>     cyl  disp  hp
#>  1:   4 120.1  97
#>  2:   8 318.0 150
#>  3:   8 304.0 150
#>  4:   8 350.0 245
#>  5:   8 400.0 175
#>  6:   4  79.0  66
#>  7:   4 120.3  91
#>  8:   4  95.1 113
#>  9:   8 351.0 264
#> 10:   6 145.0 175
#> 11:   8 301.0 335
#> 12:   4 121.0 109
#> 
#> attr(,"class")
#> [1] "glex"           "rpf_components" "list"          

# sums to prediction
cbind(
  m_sum = rowSums(components$m) + components$intercept,
  prediction = predict(rpfit, test)
)
#>       m_sum    .pred
#> 1  24.92369 24.92369
#> 2  18.04169 18.04169
#> 3  18.04169 18.04169
#> 4  14.90805 14.90805
#> 5  18.12735 18.12735
#> 6  30.77883 30.77883
#> 7  28.45402 28.45402
#> 8  26.17159 26.17159
#> 9  14.90805 14.90805
#> 10 20.78145 20.78145
#> 11 14.90805 14.90805
#> 12 24.64148 24.64148

# Only get components with interactions of a lower degree, ignoring 3-way interactions
predict_components(rpfit, test, max_interaction = 2)
#> $m
#>            cyl      disp         hp    cyl:disp       cyl:hp     disp:hp
#>  1:  1.3770452  2.603519  0.8918936  0.23153106 -0.267467990 -0.16388358
#>  2: -0.7695683 -1.465990 -0.1106307  0.01146022  0.006589163  0.23806033
#>  3: -0.7695683 -1.465990 -0.1106307  0.01146022  0.006589163  0.23806033
#>  4: -0.7695683 -1.465990 -2.8044495  0.01146022 -0.105503856 -0.26797812
#>  5: -0.7695683 -1.334206 -0.3514101  0.01096113 -0.010321204  0.42309626
#>  6:  1.3770452  4.125237  3.7886809  0.46853941 -0.002318937  0.94645152
#>  7:  1.3770452  2.603519  3.7886809  0.23153106 -0.002318937  0.31230241
#>  8:  1.3770452  4.125237  0.4823373  0.46853941 -0.205562593 -0.19865944
#>  9: -0.7695683 -1.465990 -2.8044495  0.01146022 -0.105503856 -0.26797812
#> 10: -0.4832547  1.694664 -0.3514101 -0.16986248  0.127061572 -0.11840583
#> 11: -0.7695683 -1.465990 -2.8044495  0.01146022 -0.105503856 -0.26797812
#> 12:  1.3770452  2.603519  0.4516721  0.23153106 -0.200179679 -0.05174859
#> 
#> $intercept
#> [1] 20.17863
#> 
#> $x
#>     cyl  disp  hp
#>  1:   4 120.1  97
#>  2:   8 318.0 150
#>  3:   8 304.0 150
#>  4:   8 350.0 245
#>  5:   8 400.0 175
#>  6:   4  79.0  66
#>  7:   4 120.3  91
#>  8:   4  95.1 113
#>  9:   8 351.0 264
#> 10:   6 145.0 175
#> 11:   8 301.0 335
#> 12:   4 121.0 109
#> 
#> $remainder
#>  [1]  0.07242737 -0.04685440 -0.04685440  0.13144921 -0.01982741 -0.10343096
#>  [7] -0.03536481 -0.05597535  0.13144921 -0.09596553  0.13144921  0.05100935
#> 
#> attr(,"class")
#> [1] "glex"           "rpf_components" "list"          

# Only retrieve main effects
(main_effects <- predict_components(rpfit, test, max_interaction = 1))
#> $m
#>            cyl      disp         hp
#>  1:  1.3770452  2.603519  0.8918936
#>  2: -0.7695683 -1.465990 -0.1106307
#>  3: -0.7695683 -1.465990 -0.1106307
#>  4: -0.7695683 -1.465990 -2.8044495
#>  5: -0.7695683 -1.334206 -0.3514101
#>  6:  1.3770452  4.125237  3.7886809
#>  7:  1.3770452  2.603519  3.7886809
#>  8:  1.3770452  4.125237  0.4823373
#>  9: -0.7695683 -1.465990 -2.8044495
#> 10: -0.4832547  1.694664 -0.3514101
#> 11: -0.7695683 -1.465990 -2.8044495
#> 12:  1.3770452  2.603519  0.4516721
#> 
#> $intercept
#> [1] 20.17863
#> 
#> $x
#>     cyl  disp  hp
#>  1:   4 120.1  97
#>  2:   8 318.0 150
#>  3:   8 304.0 150
#>  4:   8 350.0 245
#>  5:   8 400.0 175
#>  6:   4  79.0  66
#>  7:   4 120.3  91
#>  8:   4  95.1 113
#>  9:   8 351.0 264
#> 10:   6 145.0 175
#> 11:   8 301.0 335
#> 12:   4 121.0 109
#> 
#> $remainder
#>  [1] -0.127393139  0.209255317  0.209255317 -0.230572546  0.403908771
#>  [6]  1.309241020  0.506149728  0.008342022 -0.230572546 -0.257172269
#> [11] -0.230572546  0.030612142
#> 
#> attr(,"class")
#> [1] "glex"           "rpf_components" "list"          

# The difference is the combined contribution of interaction effects
cbind(
  m_sum = rowSums(main_effects$m) + main_effects$intercept,
  prediction = predict(rpfit, test)
)
#>       m_sum    .pred
#> 1  25.05109 24.92369
#> 2  17.83244 18.04169
#> 3  17.83244 18.04169
#> 4  15.13862 14.90805
#> 5  17.72344 18.12735
#> 6  29.46959 30.77883
#> 7  27.94787 28.45402
#> 8  26.16325 26.17159
#> 9  15.13862 14.90805
#> 10 21.03863 20.78145
#> 11 15.13862 14.90805
#> 12 24.61086 24.64148